# contracts/openapi/wiki-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nova Haven Content API",
    "version": "1.0.0",
    "description": "Draft/public isolation for Wiki, News and Catalog: anonymous clients only see published revisions. Editorial operations require Admin cookie + CSRF; mutation preconditions use If-Match."
  },
  "paths": {
...
    "/api/v1/knowledge": {
      "get": {
        "tags": ["Knowledge"],
        "parameters": [
          {"name": "kind", "in": "query", "schema": {"$ref": "#/components/schemas/KnowledgeKind"}},
          {"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}
        ],
        "responses": {"200": {"description": "Published NPC, quest, location and season snapshots only", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/KnowledgePage"}}}}, "400": {"description": "Invalid kind or pagination"}}
      }
    },
    "/api/v1/knowledge/{kind}": {
      "get": {
        "tags": ["Knowledge"],
        "parameters": [
          {"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/KnowledgeKind"}},
          {"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}
        ],
        "responses": {"200": {"description": "Published records of the selected kind", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/KnowledgePage"}}}}, "400": {"description": "Invalid kind or pagination"}}
      }
    },
    "/api/v1/knowledge/{kind}/{slug}": {
      "get": {
        "tags": ["Knowledge"],
        "parameters": [
          {"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/KnowledgeKind"}},
          {"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "Published typed record with relation snapshots"}, "404": {"description": "Unknown or unpublished record"}}
      }
    },
    "/api/v1/admin/knowledge/{kind}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/KnowledgeKind"}}], "responses": {"200": {"description": "Draft records of the selected kind"}, "400": {"description": "Invalid kind"}}},
      "post": {
        "tags": ["Admin"], "security": [{"adminCookie": []}],
        "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/KnowledgeKind"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/KnowledgeInput"}}}},
        "responses": {"201": {"description": "Draft created"}, "400": {"description": "Validation"}, "409": {"description": "Duplicate slug"}}
      }
    },
    "/api/v1/admin/knowledge/{kind}/{id}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/KnowledgeKind"}}, {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Draft with typed metadata and ETag"}, "404": {"description": "Unknown record"}}},
      "patch": {
        "tags": ["Admin"], "security": [{"adminCookie": []}],
        "parameters": [{"name": "kind", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/KnowledgeKind"}}, {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/KnowledgeInput"}}}},
        "responses": {"200": {"description": "Updated with ETag"}, "400": {"description": "Validation"}, "409": {"description": "Duplicate or invalid reference"}, "412": {"description": "Stale ETag"}, "428": {"description": "Missing ETag"}}
      }
    },
...
    "/api/v1/rewards": {
      "get": {"tags": ["Rewards"], "parameters": [{"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 100}}, {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}}, {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}], "responses": {"200": {"description": "Published reward definitions requiring external acknowledgement"}}}
    },
    "/api/v1/rewards/{slug}": {
      "get": {"tags": ["Rewards"], "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Published reward definition requiring external acknowledgement"}, "404": {"description": "Unknown or unpublished definition"}}}
    },
...
    "/api/v1/admin/rewards/{id}": {
      "get": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Reward draft with ETag"}}},
      "patch": {"tags": ["Admin"], "security": [{"adminCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RewardInput"}}}}, "responses": {"200": {"description": "Updated with ETag"}, "412": {"description": "Stale definition"}, "428": {"description": "Missing ETag"}}}
    },
...
    "/api/v1/notifications": {"get": {"tags": ["Notifications"], "security": [{"userCookie": []}], "parameters": [{"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1}}, {"name": "pageSize", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50}}], "responses": {"200": {"description": "Current player's paginated notification inbox, total and unread count"}, "400": {"description": "Invalid pagination or page out of range"}, "401": {"description": "Authentication required"}}}},
    "/api/v1/notifications/unread-count": {"get": {"tags": ["Notifications"], "security": [{"userCookie": []}], "responses": {"200": {"description": "Current player's unread notification count"}, "401": {"description": "Authentication required"}}}},
    "/api/v1/notifications/{id}/read": {"put": {"tags": ["Notifications"], "security": [{"userCookie": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"204": {"description": "Marked read"}, "404": {"description": "Unknown notification"}}}},
    "/api/v1/notifications/read-all": {"post": {"tags": ["Notifications"], "security": [{"userCookie": []}], "responses": {"204": {"description": "All current player's notifications marked read"}}}},
    "/api/v1/notifications/push/config": {"get": {"tags": ["Notifications"], "security": [{"userCookie": []}], "responses": {"200": {"description": "VAPID public key or null when disabled"}}}},
    "/api/v1/notifications/push/subscriptions": {"put": {"tags": ["Notifications"], "security": [{"userCookie": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["endpoint", "keys"], "properties": {"endpoint": {"type": "string", "format": "uri"}, "keys": {"type": "object", "required": ["p256dh", "auth"], "properties": {"p256dh": {"type": "string"}, "auth": {"type": "string"}}}}}}}}, "responses": {"204": {"description": "Push subscription saved"}, "400": {"description": "Unsupported endpoint"}, "503": {"description": "Push disabled"}}}, "delete": {"tags": ["Notifications"], "security": [{"userCookie": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["endpoint"], "properties": {"endpoint": {"type": "string", "format": "uri"}}}}}}, "responses": {"204": {"description": "Push subscription removed"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "adminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": ".AspNetCore.Identity.Application"
      },
      "userCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": ".AspNetCore.Identity.Application"
      }
    },
    "schemas": {
...
      "KnowledgeSummary": {
        "type": "object", "required": ["id", "slug", "name", "summary", "kind", "revision", "publishedAt", "previewImageUrl"],
        "properties": {"id": {"type": "string", "format": "uuid"}, "slug": {"type": "string"}, "name": {"type": "string"}, "summary": {"type": "string"}, "kind": {"$ref": "#/components/schemas/KnowledgeKind"}, "revision": {"type": "integer"}, "publishedAt": {"type": "string", "format": "date-time"}, "previewImageUrl": {"type": ["string", "null"], "maxLength": 500}}
      },
      "KnowledgePage": {
        "type": "object", "required": ["items", "page", "pageSize", "total"],
        "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/KnowledgeSummary"}}, "page": {"type": "integer"}, "pageSize": {"type": "integer"}, "total": {"type": "integer"}}
      },
...
      "KnowledgeKind": {
        "type": "string",
        "enum": ["npc", "quest", "location", "season"]
      },
      "KnowledgeLinkType": {
        "type": "string",
        "enum": ["related", "givesQuest", "locatedAt", "usesItem", "partOfSeason"]
      },
      "KnowledgeStep": {
        "type": "object",
        "additionalProperties": false,
        "required": ["position", "title", "description"],
        "properties": {
          "position": {"type": "integer", "minimum": 1, "maximum": 50},
          "title": {"type": "string", "minLength": 1, "maxLength": 160},
          "description": {"type": "string", "minLength": 1, "maxLength": 2000}
        }
      },
      "KnowledgeLink": {
        "type": "object",
        "additionalProperties": false,
        "required": ["linkType", "sortOrder"],
        "properties": {
          "linkType": {"$ref": "#/components/schemas/KnowledgeLinkType"},
          "targetEntryId": {"type": ["string", "null"], "format": "uuid"},
          "targetCatalogItemId": {"type": ["string", "null"], "format": "uuid"},
          "sortOrder": {"type": "integer", "minimum": 0, "maximum": 1000}
        }
      },
      "KnowledgeInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "slug", "summary", "markdown"],
        "properties": {
          "name": {"type": "string", "minLength": 2, "maxLength": 160},
          "slug": {"type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "minLength": 2, "maxLength": 160},
          "summary": {"type": "string", "maxLength": 500},
          "markdown": {"type": "string", "minLength": 1, "maxLength": 100000},
          "kind": {"$ref": "#/components/schemas/KnowledgeKind"},
          "role": {"type": "string", "maxLength": 120},
          "locationEntryId": {"type": ["string", "null"], "format": "uuid"},
          "portraitUrl": {"type": ["string", "null"], "maxLength": 500},
          "difficulty": {"type": "integer", "minimum": 1, "maximum": 5},
          "giverNpcEntryId": {"type": ["string", "null"], "format": "uuid"},
          "rewardDescription": {"type": "string", "maxLength": 500},
          "steps": {"type": "array", "minItems": 1, "maxItems": 50, "items": {"$ref": "#/components/schemas/KnowledgeStep"}},
          "region": {"type": "string", "maxLength": 120},
          "locationType": {"type": "string", "maxLength": 80},
          "latitude": {"type": ["number", "null"], "minimum": -90, "maximum": 90},
          "longitude": {"type": ["number", "null"], "minimum": -180, "maximum": 180},
          "mapImageUrl": {"type": ["string", "null"], "maxLength": 500},
          "startsAt": {"type": ["string", "null"], "format": "date-time"},
          "endsAt": {"type": ["string", "null"], "format": "date-time"},
          "theme": {"type": "string", "maxLength": 160},
          "eventDescription": {"type": ["string", "null"], "maxLength": 2000},
          "links": {"type": "array", "maxItems": 50, "items": {"$ref": "#/components/schemas/KnowledgeLink"}}
        }
      },
...
      "RewardKind": {"type": "string", "enum": ["item", "currency", "permission", "title", "cosmetic", "custom"]},
      "RewardInput": {"type": "object", "additionalProperties": false, "required": ["name", "slug", "summary", "markdown", "kind", "deliveryDescription"], "properties": {"name": {"type": "string", "maxLength": 160}, "slug": {"type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120}, "summary": {"type": "string", "maxLength": 500}, "markdown": {"type": "string", "maxLength": 50000}, "kind": {"$ref": "#/components/schemas/RewardKind"}, "deliveryDescription": {"type": "string", "maxLength": 1000}}},
...
    }
  }
}